Extract DefaultUnauthorizedView
