+++ packages/ra-ui-materialui/src/detail/Edit.spec.tsx
@@ -816,7 +816,6 @@ describe('<Edit />', () => {
             await waitFor(() => {
                 expect(screen.getAllByText('Hello')).toHaveLength(1);
             });
-            expect(screen.queryAllByText('Hello')).toHaveLength(1);
         });
     });
 
@@ -904,9 +903,11 @@ describe('<Edit />', () => {
     describe('Access control', () => {
         it('should display the edit page when the user has access to the resource', async () => {
             render(<AccessControl />);
+            // First test that we have access to the resource
             await screen.findByDisplayValue('War and Peace');
+            // Then test that we react to the access control change
             fireEvent.click(await screen.findByLabelText('books access'));
-            await screen.findByText('ra-rbac.page.unauthorized');
+            await screen.findByText('ra.page.unauthorized');
         });
     });
 });
