+++ packages/ra-ui-materialui/src/detail/EditView.tsx
@@ -1,18 +1,12 @@
 import * as React from 'react';
 import { ReactElement, ElementType } from 'react';
-import { Box, Button, Card, CardContent, styled, SxProps } from '@mui/material';
+import { Card, CardContent, styled, SxProps } from '@mui/material';
 import clsx from 'clsx';
-import {
-    useEditContext,
-    useResourceDefinition,
-    useCanAccess,
-    useTranslate,
-} from 'ra-core';
-import LockIcon from '@mui/icons-material/Lock';
+import { useEditContext, useResourceDefinition, useCanAccess } from 'ra-core';
 
 import { EditActions } from './EditActions';
 import { Title } from '../layout';
-import { History } from '@mui/icons-material';
+import { DefaultUnauthorizedView } from './DefaultUnauthorizedView';
 
 export const EditView = (props: EditViewProps) => {
     const {
@@ -22,7 +16,7 @@ export const EditView = (props: EditViewProps) => {
         className,
         component: Content = Card,
         title,
-        unauthorizedView = <DefaultUnauthorizedView />,
+        unauthorized = <DefaultUnauthorizedView />,
         ...rest
     } = props;
 
@@ -42,7 +36,7 @@ export const EditView = (props: EditViewProps) => {
     }
 
     if (!canAccess) {
-        return unauthorizedView;
+        return unauthorized;
     }
 
     return (
@@ -76,62 +70,10 @@ export interface EditViewProps
     component?: ElementType;
     title?: string | ReactElement | false;
     sx?: SxProps;
-    unauthorizedView?: ReactElement;
+    unauthorized?: ReactElement;
 }
 
 const defaultActions = <EditActions />;
-
-function goBack() {
-    window.history.go(-1);
-}
-
-const DefaultUnauthorizedView = () => {
-    const translate = useTranslate();
-    return (
-        <Box
-            sx={theme => ({
-                display: 'flex',
-                flexDirection: 'column',
-                alignItems: 'center',
-                justifyContent: 'center',
-                [theme.breakpoints.down('md')]: {
-                    padding: '1em',
-                },
-                fontFamily: 'Roboto, sans-serif',
-                opacity: 0.5,
-                '& h1': {
-                    display: 'flex',
-                    alignItems: 'center',
-                },
-            })}
-        >
-            <h1 role="alert">
-                <LockIcon
-                    sx={{
-                        width: '2em',
-                        height: '2em',
-                        marginRight: '0.5em',
-                    }}
-                />
-                {translate('ra-rbac.page.unauthorized')}
-            </h1>
-            <div>{translate('ra-rbac.message.unauthorized')}</div>
-            <Box
-                sx={{
-                    marginTop: '2em',
-                }}
-            >
-                <Button
-                    variant="contained"
-                    startIcon={<History />}
-                    onClick={goBack}
-                >
-                    {translate('ra.action.back')}
-                </Button>
-            </Box>
-        </Box>
-    );
-};
 
 const PREFIX = 'RaEdit';
